chore(*): Update Naming Conventions
* Update file names to be closer to [angular.io style
* guide](https://angular.io/styleguide)
* Change index.ts -> thing.component.ts
* Change index.test.ts -> thing.component.test.ts
* Add index.ts where appropiate and export from thing.component.ts
* Update test files to import from correct location
