+++ src/components/button/button.component.ts
@@ -1,0 +1,23 @@
+import { Component, EventEmitter, Input, Output } from '@angular/core';
+
+@Component({
+  selector: 'rio-button',
+  template: `
+    <button
+      (click)="handleClick($event)"
+      type="{{type || 'button'}}"
+      class="btn btn-primary {{className}}">
+
+      <ng-content></ng-content>
+    </button>
+  `
+})
+export class RioButtonComponent {
+  @Input() className: string;
+  @Input() type: string;
+  @Output() onClick: EventEmitter<any> = new EventEmitter<any>();
+
+  handleClick(event: any) {
+    this.onClick.emit(event);
+  }
+};
