+++ src/components/counter/counter.component.test.ts
