+++ src/components/logo/logo.component.test.ts
