+++ src/components/navigator/index.ts
@@ -1,4 +1,4 @@
-import { RioNavigatorComponent } from './navigator';
-import { RioNavigatorItemComponent } from './navigator-item';
+import { RioNavigatorComponent } from './navigator.component';
+import { RioNavigatorItemComponent } from './navigator-item.component';
 
 export { RioNavigatorComponent, RioNavigatorItemComponent };
